TBB 3.0 update 7
These are the download files for the 20110427 stable release. These are
the same sources/binaries as in tbb30_196oss comm-aligned release.

+++ examples/task_priority/fractal/fractal.cpp
@@ -79,7 +79,7 @@ color_t fractal::calc_one_pixel(int x0, int y0) {
         iter++;
     }
 
-    int b = 256*mu;
+    int b = (int)(256*mu);
     int g = (b/8);
     int r = (g/16);
